+++ libs/designer-ui/src/lib/settings/settingsection/settingTokenField.tsx
@@ -1,5 +1,6 @@
 import { ArrayEditor } from '../../arrayeditor';
 import { AuthenticationEditor } from '../../authentication';
+import type { FileNameChangeHandler } from '../../code';
 import { CodeEditor } from '../../code';
 import { isCustomCode } from '../../code/util';
 import { Combobox } from '../../combobox';
@@ -12,12 +13,12 @@ import type {
   CallbackHandler,
   CastHandler,
   ChangeHandler,
-  FileNameChangeHandler,
   GetTokenPickerHandler,
   loadParameterValueFromStringHandler,
 } from '../../editor/base';
 import type { TokenPickerButtonEditorProps } from '../../editor/base/plugins/tokenpickerbutton';
 import { createLiteralValueSegment, getDropdownOptionsFromOptions } from '../../editor/base/utils/helper';
+import { InitializeVariableEditor } from '../../editor/initializevariable';
 import { StringEditor } from '../../editor/string';
 import { FloatingActionMenuKind } from '../../floatingactionmenu/constants';
 import { FloatingActionMenuInputs } from '../../floatingactionmenu/floatingactionmenuinputs';
@@ -38,8 +39,13 @@ import { CustomTokenField, isCustomEditor } from './customTokenField';
 import { Label } from '../../label';
 import { EditorLanguage, equals, getPropertyValue, replaceWhiteSpaceWithUnderscore } from '@microsoft/logic-apps-shared';
 import { MixedInputEditor } from '../../mixedinputeditor/mixedinputeditor';
-import { useMemo } from 'react';
+import { cloneElement, useMemo, useState } from 'react';
 import { useIntl } from 'react-intl';
+
+interface EditorHostOptions {
+  suppressCastingForSerialize?: boolean;
+  isMultiVariableEnabled?: boolean;
+}
 
 export interface SettingTokenFieldProps extends SettingProps {
   id?: string;
@@ -70,7 +76,9 @@ export interface SettingTokenFieldProps extends SettingProps {
   getTokenPicker: GetTokenPickerHandler;
   validationErrors?: string[];
   hideValidationErrors?: ChangeHandler;
-  suppressCastingForSerialize?: boolean;
+  hostOptions?: EditorHostOptions;
+  subComponent?: JSX.Element | null;
+  subMenu?: JSX.Element | null;
 }
 
 export const SettingTokenField = ({ ...props }: SettingTokenFieldProps) => {
@@ -79,16 +87,24 @@ export const SettingTokenField = ({ ...props }: SettingTokenFieldProps) => {
   const hideLabel =
     (isCustomEditor(props) && props.editorOptions?.hideLabel === true) ||
     equals(props.editor?.toLowerCase(), constants.PARAMETER.EDITOR.FLOATINGACTIONMENU);
+  const [showSubComponent, setShowSubComponent] = useState(false);
+
   return (
     <>
       {!hideLabel && (
-        <div className="msla-input-parameter-label">
+        <div className="msla-input-parameter-label" style={{ display: 'flex', justifyContent: 'space-between' }}>
           <Label id={labelId} isRequiredField={props.required} text={props.label} />
+          {props.subMenu && cloneElement(props.subMenu, { setShowSubComponent })}
         </div>
       )}
       <div key={props.id}>
         {isCustomEditor(props) ? <CustomTokenField {...props} labelId={labelId} /> : <TokenField {...props} labelId={labelId} />}
       </div>
+      {props.subComponent ? (
+        <div className="msla-input-parameter-subcomponent">
+          {cloneElement(props.subComponent, { showSubComponent, setShowSubComponent })}
+        </div>
+      ) : null}
     </>
   );
 };
@@ -114,11 +130,11 @@ export const TokenField = ({
   loadParameterValueFromString,
   onValueChange,
   onComboboxMenuOpen,
+  onFileNameChange,
   hideValidationErrors,
   onCastParameter,
-  onFileNameChange,
   getTokenPicker,
-  suppressCastingForSerialize,
+  hostOptions,
   required,
 }: TokenFieldProps) => {
   const intl = useIntl();
@@ -166,7 +182,7 @@ export const TokenField = ({
           onMenuOpen={onComboboxMenuOpen}
           tokenMapping={tokenMapping}
           loadParameterValueFromString={loadParameterValueFromString}
-          suppressCastingForSerialize={suppressCastingForSerialize}
+          suppressCastingForSerialize={hostOptions?.suppressCastingForSerialize}
         />
       );
 
@@ -365,6 +381,26 @@ export const TokenField = ({
           getTokenPicker={getTokenPicker}
           onChange={onValueChange}
           dataAutomationId={`msla-setting-token-editor-htmleditor-${labelForAutomationId}`}
+        />
+      );
+
+    case constants.PARAMETER.EDITOR.INITIALIZE_VARIABLE:
+      return (
+        <InitializeVariableEditor
+          initialValue={value}
+          getTokenPicker={getTokenPicker}
+          onChange={(updatedChangeState) => {
+            onValueChange?.(updatedChangeState);
+            hideValidationErrors?.(updatedChangeState);
+          }}
+          validationErrors={editorViewModel.validationErrors}
+          tokenMapping={tokenMapping}
+          loadParameterValueFromString={loadParameterValueFromString}
+          readonly={readOnly}
+          isAgentParameter={editorOptions?.isAgentParameter}
+          tokenPickerButtonProps={tokenpickerButtonProps}
+          dataAutomationId={`msla-setting-token-editor-initializevariableeditor-${labelForAutomationId}`}
+          isMultiVariableEnabled={hostOptions?.isMultiVariableEnabled}
         />
       );